HUD added
self-made HUD to handle errors

+++ Sibers Image Processor/Classes/ImageProcessor/SIPImageProcessorDelegate.h
@@ -9,6 +9,6 @@
 @protocol SIPImageProcessorDelegate <NSObject>
 
 - (void)imageProcessorDoneProcessingImage: (UIImage *)image;
-- (void)imageProcessorProcessingImage: (CGFloat)progress;
+- (void)imageProcessorProcessingImage: (float)progress;
 
 @end
